Fix false positive call on mocking is_on method
Fix #553

Previsouly, the tests/test_light was patching the is_on method and the
real method is never executed, and thus the test pass but not hit the
tested code.

This is classic situation where what we want to mock isnt the thing, but
the dependencies of the thing (in this case, what is beign called inside
what we want to test).

Instead mock the is_on method, is necessary to mock the
IOCircuit::circuit.value() call (that is really what we want to test).

This commit changes the toggles in mocker from ligth.is_on to
light.circuit.value.