fix: link policy_type to script_type categories in `DefaultWallet`
Create a new class `LinkedCategorySetting` so it can add a relation
between two categories (in this case only for a relation between wallet
policy and the proper script type).

Before this commit the firmware was using cosigner 0 as default for
p2sh. This commit add some checks for it and apply partial scripts fix
for multisig keys when using p2sh.

If the user set the default policy as multisig and default script to
p2sh, the default created key will be set with a `None` account.
Otherwise, the default account will be zero.

Also some POLICY and SCRIPT constants (that define the labels) was moved
from settings to key module to avoid circular importing due to the added
`LinkedCategorySetting`.

Co-authored-by: odudex <odudex@proton.me>