fix(xor): low entropy calc when using Mnemonic XOR
This commit add a fix for shannon entropy validation. Before this commit
we checked for only two specific cases (zero/all-ff entropy).

We can, instead, use a factor that measures the ammount of information
have this entropy,  made by the reason between a custom threshold and
the current word 11 count. With this reason, we can add a rationale
use decision if this or another mnemonic could be what could have a
"dangerous entropy for the current context: entropies that have less
information than the current mnemonic entropy.

This can be tunned by a constant called `MIN_BITS_PER_WORD11` that is a
float value measuring the amount of entropy of the involved mnemonics.