ci: add convetional commits to krux
Convetional commits are a specification for adding human and machine
readable meaning to commit messages. This commit checks if the follow-up
commits agree with those specified in the
https://www.conventionalcommits.org/en/v1.0.0/. While this changes some
of the current workflow for commits, this brings more organization and a
development standard (the semmantic git commit message) to a project
that had a lot of attention in the last months.