feat: add signet and regtest networks
This commit add the options `signet` and `regtest` to `DefaultWallet` at
`src/krux/krux_settings` as well `Signet` and `Regtest` names to
`src/krux/wallet_settings`. To be in coherent with the change, it was
added the options string to menubar as well added theme colors for it.

It's worth to note that regtest and signet could be added to test
vectors as well used to create new vector for integration tests.